+++ src/entities/quiz/model/types.ts
@@ -1,0 +1,8 @@
+import type { components } from '#/shared/api'
+
+export type QuizBrief = components['schemas']['QuizBrief']
+export type QuizDetail = components['schemas']['QuizDetail']
+export type Question = components['schemas']['QuestionRead']
+export type Answer = components['schemas']['AnswerRead']
+export type QuizResult = components['schemas']['QuizResultRead']
+export type QuizStartResult = components['schemas']['QuizStartResult']